File Exist Attempt
- formatted date and time to be human readable
- attempted to create if/else condition for if file already exists
- attempt failed <<< HELP >>>

+++ Jonny_.py
@@ -5,23 +5,26 @@
 
 import os
 import datetime
+import os.path
 
 path = os.path.expanduser("~/Desktop")
 os.chdir(path)
 
-#Time needs to be converted into a string, research 'strftime' and tell me what it does.
-currentDT = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
-
-# 'strftime' is a method that basically takes 'datetime' and turn it into a string to make it human readable
-# 'strptime' does the opposite
-
 newfolder = os.mkdir("Wakanda Forever")
-
 path = os.path.expanduser("~/Desktop/Wakanda Forever")
 
 os.chdir(path)
+os.path.exists(path)
 
-f= open("index.html", "w+")
+currentDT = datetime.datetime.now().strftime("%m/%d/%Y - %H:%M")
 
-f.write(currentDT)
-f.close()
+if os.path.exists(path):
+    f= open("index.html", "w+")
+    f.write(currentDT)
+    f.close()
+
+else:
+    os.chdir(path)
+    f= open("index.html", "w+")
+    f.write(currentDT)
+    f.close()
